requirements.txt
beautifulsoup4
python-telegram-bot
python-dateutil
requests
PyYAML